linkedin_service/requirements.txt
scrapfly-sdk[all]
jmespath
loguru
parsel
flask
python-dotenv
